Enum dbus::arg::messageitem::MessageItem

pub enum MessageItem {

    Array(MessageItemArray),
    Struct(Vec<MessageItem>),
    Variant(Box<MessageItem>),
    Dict(MessageItemDict),
    ObjectPath(Path<'static>),
    Signature(Signature<'static>),
    Str(String),
    Bool(bool),
    Byte(u8),
    Int16(i16),
    Int32(i32),
    Int64(i64),
    UInt16(u16),
    UInt32(u32),
    UInt64(u64),
    Double(f64),
    UnixFd(OwnedFd),

}

MessageItem - used as parameters and return values from method calls, or as data added to a signal (old, enum version).

Note that the newer generic design (see arg module) is both faster and less error prone than MessageItem, and should be your first hand choice whenever applicable.

Variants

Array(MessageItemArray)

A D-Bus array requires all elements to be of the same type. All elements must match the Signature.

Tuple Fields of Array

0: MessageItemArray

Struct(Vec<MessageItem>)

A D-Bus struct allows for values of different types.

Tuple Fields of Struct

0: Vec<MessageItem>

Variant(Box<MessageItem>)

A D-Bus variant is a wrapper around another MessageItem, which can be of any type.

Tuple Fields of Variant

0: Box<MessageItem>

Dict(MessageItemDict)

A D-Bus dictionary. All keys and values are required to be of the same type. Not all types can be dictionary keys, but all can be dictionary values.

Tuple Fields of Dict

0: MessageItemDict

ObjectPath(Path<'static>)

A D-Bus objectpath requires its content to be a valid objectpath, so this cannot be any string.

Tuple Fields of ObjectPath

0: Path<'static>

Signature(Signature<'static>)

A D-Bus signature requires its content to be a valid type signature, so this cannot be any string.

Tuple Fields of Signature

0: Signature<'static>

Str(String)

A D-Bus String is zero terminated, so no \0 s in the String, please. (D-Bus strings are also - like Rust strings - required to be valid UTF-8.)

Tuple Fields of Str

0: String

Bool(bool)

A D-Bus boolean type.

Tuple Fields of Bool

0: bool

Byte(u8)

A D-Bus unsigned 8 bit type.

Tuple Fields of Byte

0: u8

Int16(i16)

A D-Bus signed 16 bit type.

Tuple Fields of Int16

0: i16

Int32(i32)

A D-Bus signed 32 bit type.

Tuple Fields of Int32

0: i32

Int64(i64)

A D-Bus signed 64 bit type.

Tuple Fields of Int64

0: i64

UInt16(u16)

A D-Bus unsigned 16 bit type.

Tuple Fields of UInt16

0: u16

UInt32(u32)

A D-Bus unsigned 32 bit type.

Tuple Fields of UInt32

0: u32

UInt64(u64)

A D-Bus unsigned 64 bit type.

Tuple Fields of UInt64

0: u64

Double(f64)

A D-Bus IEEE-754 double-precision floating point type.

Tuple Fields of Double

0: f64

UnixFd(OwnedFd)

D-Bus allows for sending file descriptors, which can be used to set up SHM, unix pipes, or other communication channels.

Tuple Fields of UnixFd

0: OwnedFd

Implementations

impl MessageItem

pub fn signature(&self) -> Signature<'static>

Get the D-Bus Signature for this MessageItem.

pub fn arg_type(&self) -> ArgType

Get the arg type of this MessageItem.

pub fn from_dict<E, I: Iterator<Item = Result<(String, MessageItem), E>>>(
i: I
) -> Result<MessageItem, E>

Creates a (String, Variant) dictionary from an iterator with Result passthrough (an Err will abort and return that Err)

pub fn new_array(v: Vec<MessageItem>) -> Result<MessageItem, ArrayError>

Creates an MessageItem::Array from a list of MessageItems.

Note: This requires v to be non-empty. See also MessageItem::from(&[T]), which can handle empty arrays as well.

pub fn new_dict(
v: Vec<(MessageItem, MessageItem)>
) -> Result<MessageItem, ArrayError>

Creates an MessageItem::Dict from a list of MessageItem pairs.

Note: This requires v to be non-empty. See also MessageItem::from(&[(T1, T2)]), which can handle empty arrays as well.

pub fn inner<'a, T: TryFrom<&'a MessageItem>>(&'a self) -> Result<T, T::Error>

Get the inner value of a MessageItem

Example

use dbus::arg::messageitem::MessageItem;
let m: MessageItem = 5i64.into();
let s: i64 = m.inner().unwrap();
assert_eq!(s, 5i64);

pub fn peel(&self) -> &Self

Get the underlying MessageItem of a MessageItem::Variant

Nested MessageItem::Variants are unwrapped recursively until a non-Variant is found.

Example

use dbus::arg::messageitem::MessageItem;
let nested = MessageItem::Variant(Box::new(6i64.into()));
let flat: MessageItem = 6i64.into();
assert_ne!(&nested, &flat);
assert_eq!(nested.peel(), &flat);