Struct dbus::arg::Iter

pub struct Iter<'a>(DBusMessageIter, &'a Message, u32);

Helper struct for retrieve one or more arguments from a Message.

Methods

impl<'a> Iter<'a>

fn new(m: &'a Message) -> Iter<'a>

Creates a new struct for iterating over the arguments of a message, starting with the first argument.

fn get<T: Get<'a>>(&mut self) -> Option<T>

Returns the current argument, if T is the argument type. Otherwise returns None.

fn get_refarg(&mut self) -> Option<Box<RefArg + 'static>>

Returns the current argument as a trait object (experimental).

Note: For the more complex arguments (arrays / dicts / structs, and especially combinations thereof), their internal representations are still a bit in flux. Instead, use as_iter() to read the values of those.

The rest are unlikely to change - Variants are Variant>, strings are String, paths are Path<'static>, signatures are Signature<'static>, Int32 are i32s and so on.

fn signature(&mut self) -> Signature<'static>

Returns the type signature for the current argument.

fn arg_type(&mut self) -> ArgType

The raw arg_type for the current item.

Unlike Arg::arg_type, this requires access to self and is not a static method. You can match this against Arg::arg_type for different types to understand what type the current item is. In case you're past the last argument, this function will return 0.

fn next(&mut self) -> bool

Returns false if there are no more items.

fn read<T: Arg + Get<'a>>(&mut self) -> Result<T, TypeMismatchError>

Wrapper around get and next. Calls get, and then next if get succeeded.

Also returns a Result rather than an Option to work better with try!.

Example

struct ServiceBrowserItemNew {
    interface: i32,
    protocol: i32,
    name: String,
    item_type: String,
    domain: String,
    flags: u32,
}

fn service_browser_item_new_msg(m: &Message) -> Result<ServiceBrowserItemNew, TypeMismatchError> {
    let mut iter = m.iter_init();
    Ok(ServiceBrowserItemNew {
        interface: try!(iter.read()),
        protocol: try!(iter.read()),
        name: try!(iter.read()),
        item_type: try!(iter.read()),
        domain: try!(iter.read()),
        flags: try!(iter.read()),
    })
}

fn recurse(&mut self, arg_type: ArgType) -> Option<Iter<'a>>

If the current argument is a container of the specified arg_type, then a new Iter is returned which is for iterating over the contents inside the container.

Primarily for internal use (the "get" function is more ergonomic), but could be useful for recursing into containers with unknown types.