Struct dbus::arg::IterAppend

pub struct IterAppend<'a>(DBusMessageIter, &'a Message);

Helper struct for appending one or more arguments to a Message.

Tuple Fields
0: DBusMessageIter1: &'a Message
Implementations

impl<'a> IterAppend<'a>

pub fn new(m: &'a mut Message) -> IterAppend<'a>

Creates a new IterAppend struct.

pub fn append<T: Append>(&mut self, a: T)

Appends the argument.

fn append_container<F: FnOnce(&mut IterAppend<'a>)>( &mut self, arg_type: ArgType, sig: Option<&CStr>, f: F )

pub fn append_variant<F: FnOnce(&mut IterAppend<'a>)>( &mut self, inner_sig: &Signature<'_>, f: F )

Low-level function to append a variant.

Use in case the Variant struct is not flexible enough - the easier way is to just call e g “append1” on a message and supply a Variant parameter.

In order not to get D-Bus errors: during the call to “f” you need to call “append” on the supplied IterAppend exactly once, and with a value which has the same signature as inner_sig.

pub fn append_array<F: FnOnce(&mut IterAppend<'a>)>( &mut self, inner_sig: &Signature<'_>, f: F )

Low-level function to append an array.

Use in case the Array struct is not flexible enough - the easier way is to just call e g “append1” on a message and supply an Array parameter.

In order not to get D-Bus errors: during the call to “f”, you should only call “append” on the supplied IterAppend with values which has the same signature as inner_sig.

pub fn append_struct<F: FnOnce(&mut IterAppend<'a>)>(&mut self, f: F)

Low-level function to append a struct.

Use in case tuples are not flexible enough - the easier way is to just call e g “append1” on a message and supply a tuple parameter.

pub fn append_dict_entry<F: FnOnce(&mut IterAppend<'a>)>(&mut self, f: F)

Low-level function to append a dict entry.

Use in case the Dict struct is not flexible enough - the easier way is to just call e g “append1” on a message and supply a Dict parameter.

In order not to get D-Bus errors: during the call to “f”, you should call “append” once for the key, then once for the value. You should only call this function for a subiterator you got from calling “append_dict”, and signatures need to match what you specified in “append_dict”.

pub fn append_dict<F: FnOnce(&mut IterAppend<'a>)>( &mut self, key_sig: &Signature<'_>, value_sig: &Signature<'_>, f: F )

Low-level function to append a dict.

Use in case the Dict struct is not flexible enough - the easier way is to just call e g “append1” on a message and supply a Dict parameter.

In order not to get D-Bus errors: during the call to “f”, you should only call “append_dict_entry” for the subiterator - do this as many times as the number of dict entries.