Module dbus::blocking

Connections and proxies that make blocking method calls.

Modules

stdintf
This module contains some standard interfaces and an easy way to call them.

Structs

Connection
A connection to D-Bus, non-async version where callbacks are Send but not Sync.

LocalConnection
A connection to D-Bus, thread local + non-async version

Proxy
A struct that wraps a connection, destination and path.

SyncConnection
A connection to D-Bus, Send + Sync + non-async version

Traits

BlockingSender
Abstraction over different connections

MakeSignal
Internal helper trait