Struct dbus::blocking::SyncConnection

pub struct SyncConnection { /* fields omitted */ }

A connection to D-Bus, Send + Sync + non-async version

Implementations

impl SyncConnection

pub fn new_session() -> Result<Self, Error>

Create a new connection to the session bus.

pub fn new_system() -> Result<Self, Error>

Create a new connection to the system-wide bus.

pub fn unique_name(&self) -> BusName

Get the connection's unique name.

It's usually something like ":1.54"

pub fn with_proxy<'a, 'b, D: Into<BusName<'a>>, P: Into<Path<'a>>>(
&'b self,
dest: D,
path: P,
timeout: Duration
) -> Proxy<'a, &'b Self>

Create a convenience struct for easier calling of many methods on the same destination and path.

pub fn request_name<'a, N: Into<BusName<'a>>>(
&self,
name: N,
allow_replacement: bool,
replace_existing: bool,
do_not_queue: bool
) -> Result<RequestNameReply, Error>

Request a name on the D-Bus.

For detailed information on the flags and return values, see the libdbus documentation.

pub fn release_name<'a, N: Into<BusName<'a>>>(
&self,
name: N
) -> Result<ReleaseNameReply, Error>

Release a previously requested name on the D-Bus.

pub fn add_match<S: ReadAll, F>(
&self,
match_rule: MatchRule<'static>,
f: F
) -> Result<Token, Error> where
F: FnMut(S, &Self, &Message) -> bool + Send + Sync + 'static,

Adds a new match to the connection, and sets up a callback when this message arrives.

The returned value can be used to remove the match. The match is also removed if the callback returns "false".

pub fn add_match_no_cb(&self, match_str: &str) -> Result<(), Error>

Adds a new match to the connection, without setting up a callback when this message arrives.

pub fn remove_match_no_cb(&self, match_str: &str) -> Result<(), Error>

Removes a match from the connection, without removing any callbacks.

pub fn remove_match(&self, id: Token) -> Result<(), Error>

Removes a previously added match and callback from the connection.

pub fn process(&self, timeout: Duration) -> Result<bool, Error>

Tries to handle an incoming message if there is one. If there isn't one, it will wait up to timeout

This method only takes "&self" instead of "&mut self", but it is a logic error to call it recursively and might lead to panics or deadlocks.

pub fn channel(&self) -> &Channel

The channel for this connection