Enum dbus::ConnectionItem

pub enum ConnectionItem {
    Nothing,
    MethodCall(Message),
    Signal(Message),
    MethodReturn(Message),
    WatchFd(Watch),
}

When listening for incoming events on the D-Bus, this enum will tell you what type of incoming event has happened.

Variants

Nothing

No event between now and timeout

MethodCall

Incoming method call

Signal

Incoming signal

MethodReturn

Incoming method return (mostly used for Async I/O)

WatchFd

Indicates whether a file descriptor should be monitored or not. Unless you're doing Async I/O, you can simply ignore this variant.