Module dbus::ffidisp

A connection that uses FFI callbacks to dispatch messages.

This is the legacy design used up to 0.6.x. It is not recommended for new development.

Modules

stdintf

This module contains some standard interfaces and an easy way to call them.

Structs

ConnMsgs

Iterator over incoming messages on a connection.

ConnPath

A convenience struct that wraps connection, destination and path.

Connection

A D-Bus connection. Start here if you want to get on the D-Bus!

MessageReply

The struct returned from Connection::send_and_reply.

MsgHandlerResult

The result from MsgHandler::handle.

Watch

A file descriptor, and an indication whether it should be read from, written to, or both.

Enums

BusType

System or Session bus

ConnectionItem

When listening for incoming events on the D-Bus, this enum will tell you what type of incoming event has happened.

MsgHandlerType

Type of messages to be handled by a MsgHandler.

NameFlag

Flags to use for Connection::register_name.

ReleaseNameReply

RequestNameReply

WatchEvent

A file descriptor to watch for incoming events (for async I/O).

Traits

MsgHandler

A trait for handling incoming messages.

Type Definitions

MessageCallback

The type of function to use for replacing the message callback.