Struct dbus::ffidisp::ConnPath

pub struct ConnPath<'a, C> {
    pub conn: C,
    pub dest: BusName<'a>,
    pub path: Path<'a>,
    pub timeout: i32,
}

A convenience struct that wraps connection, destination and path.

Useful if you want to make many method calls to the same destination path.

Fields
conn: C

Some way to access the connection, e g a &Connection or Rc

dest: BusName<'a>

Destination, i e what D-Bus service you’re communicating with

path: Path<'a>

Object path on the destination

timeout: i32

Timeout in milliseconds for blocking method calls

Implementations

impl<'a, C: Deref<Target = Connection>> ConnPath<'a, C>

pub fn method_call_with_args<F: FnOnce(&mut Message)>(
&self,
i: &Interface<'_>,
m: &Member<'_>,
f: F
) -> Result<Message, Error>

Make a D-Bus method call, where you can append arguments inside the closure.

pub fn signal_with_args<F: FnOnce(&mut Message)>(
&self,
i: &Interface<'_>,
m: &Member<'_>,
f: F
) -> Result<u32, Error>

Emit a D-Bus signal, where you can append arguments inside the closure.

pub fn emit<S: SignalArgs + AppendAll>(&self, signal: &S) -> Result<u32, Error>

Emit a D-Bus signal, where the arguments are in a struct.

pub fn method_call<'i, 'm, R: ReadAll, A: AppendAll, I: Into<Interface<'i>>, M: Into<Member<'m>>>(
&self,
i: I,
m: M,
args: A
) -> Result<R, Error>

Make a method call using typed input and output arguments.

Example

use dbus::ffidisp::{Connection, BusType};

let conn = Connection::get_private(BusType::Session)?;
let dest = conn.with_path("org.freedesktop.DBus", "/", 5000);
let (has_owner,): (bool,) = dest.method_call("org.freedesktop.DBus", "NameHasOwner", ("dummy.name.without.owner",))?;
assert_eq!(has_owner, false);