Module dbus::methoddisp

Structs

Argument
A D-Bus Argument.

Factory
The factory is used to create object paths, interfaces, methods etc.

Interface
Represents a D-Bus interface.

Method
A D-Bus Method.

MethodErr
A D-Bus Method Error.

MethodFn
A MethodType that wraps an Fn function

MethodFnMut
A MethodType that wraps an FnMut function. Calling this recursively will cause a refcell panic.

MethodSync
A MethodType that wraps an Fn+Send+Sync function, so it can be called from several threads in parallel.

ObjectPath
A D-Bus Object Path.

Property
A D-Bus Property.

Signal
A D-Bus Signal.

Tree
A collection of object paths.

TreeServer
An iterator adapter that handles incoming method calls.

Enums

Access
The possible access characteristics a Property can have.

EmitsChangedSignal
Enumerates the different signaling behaviors a Property can have to being changed.

Traits

MethodType
A helper trait used internally to make the tree generic over MethodFn, MethodFnMut and MethodSync.

Functions

introspect_anns

introspect_map

Type Definitions

ArcMap

MethodResult
Result containing the Messages returned from the Method, or a MethodErr.