Module dbus::msgarg

Structs

Array
Represents a D-Bus Array. Maximum flexibility (wraps an iterator of items to append). Note: Slices of FixedArray can be faster.

Dict
Append a D-Bus dict type (i e, an array of dict entries).

Iter
Helper struct for retrieve one or more arguments from a Message. Note that this is not a Rust iterator, because arguments are often of different types

IterAppend
Helper struct for appending one or more arguments to a Message.

Variant
A simple wrapper to specify a D-Bus variant.

Constants

FIXED_ARRAY_ALIGNMENTS

Traits

Append
Types that can be appended to a message as arguments implement this trait.

Arg
Types that can represent a D-Bus message argument implement this trait.

DictKey
Types that can be used as keys in a dict type implement this trait.

FixedArray
If a type implements this trait, it means the size and alignment is the same as in D-Bus. This means that you can quickly append and get slices of this type.

Get
Types that can be retrieved from a message as arguments implement this trait.

Functions

arg_append_basic

arg_append_f64

arg_append_str

arg_get_basic

arg_get_f64

arg_get_str

check

ffi_iter

test_compile