Struct dbus::obj::ObjectPath

pub struct ObjectPath<'a> {
    // some fields omitted
}

Represents a D-Bus object path, which can in turn contain Interfaces.

Methods

impl<'a> ObjectPath<'a>

fn new(conn: &'a Connection, path: &str, introspectable: bool) -> ObjectPath<'a>

Create a new ObjectPath.

fn insert_interface<N: ToString>(&mut self, name: N, i: Interface<'a>)

Add an Interface to this ObjectPath.

fn is_registered(&self) -> bool

Returns if the ObjectPath is registered.

fn set_registered(&mut self, register: bool) -> Result<(), Error>

Changes the registration status of the ObjectPath.

fn handle_message(&mut self, msg: &mut Message) -> Option<Result<(), ()>>

Handles a method call if the object path matches. Return value: None => not handled (no match), Some(Err(())) => message reply send failed, Some(Ok()) => message reply send ok */