Module dbus::objpath

Structs

Annotation

Argument
a Method has a list of Arguments.

IMethod

IObjectPath

IProperty

ISignal

Interface
Interfaces can contain Methods, Properties, and Signals.

Method
a method that can be called from another application

ObjectPath
Represents a D-Bus object path, which can in turn contain Interfaces.

Property
Properties that a remote application can get/set.

Signal
Declares that an Interface can send this signal

Enums

PropertyAccess
Types of access to a Property.

Traits

PropertyROHandler
A read-only property handler.

PropertyRWHandler
A read/write property handler.

PropertyWOHandler
A write-only property handler.

Functions

introspect_anns

introspect_args

introspect_map

parse_msg_str

parse_msg_variant

Type Definitions

MethodHandler
A boxed closure for dynamic dispatch. It is called when the method is called by a remote application.

MethodResult
A method returns either a list of MessageItems, or an error - the tuple represents the name and message of the Error.

PropertyGetResult
Contains the retrieved MessageItem or an error tuple containing the name and message of the error.

PropertySetResult
Contains () or an error tuple containing the name and message of the error.