Trait dbus::signalargs::SignalArgs

pub trait SignalArgs: Default {
    const NAME: &'static str;
    const INTERFACE: &'static str;

    // Required methods
    fn append(&self, i: &mut IterAppend<'_>);
    fn get(&mut self, i: &mut Iter<'_>) -> Result<(), TypeMismatchError>;

    // Provided methods
    fn to_emit_message(&self, path: &Path<'_>) -> Message { ... }
    fn from_message(m: &Message) -> Option<Self> { ... }
    fn match_rule<'a>(
        sender: Option<&'a BusName<'_>>,
        path: Option<&'a Path<'_>>
    ) -> MatchRule<'a> { ... }
    fn match_str(
        sender: Option<&BusName<'_>>,
        path: Option<&Path<'_>>
    ) -> String { ... }
}

Helper methods for structs representing a Signal

Example

Listen to InterfacesRemoved signal from org.bluez.obex.

use dbus::{Connection, BusType, SignalArgs};
use dbus::stdintf::org_freedesktop_dbus::ObjectManagerInterfacesRemoved as IR;

let c = Connection::get_private(BusType::Session).unwrap();
// Add a match for this signal
let mstr = IR::match_str(Some(&"org.bluez.obex".into()), None);
c.add_match(&mstr).unwrap();

// Wait for the signal to arrive.
for msg in c.incoming(1000) {
    if let Some(ir) = IR::from_message(&msg) {
        println!("Interfaces {:?} have been removed from bluez on path {}.", ir.interfaces, ir.object);
    }
}

Required Associated Constants

const NAME: &'static str

D-Bus name of signal

const INTERFACE: &'static str

D-Bus name of interface this signal belongs to

Required Methods

fn append(&self, i: &mut IterAppend<'_>)

Low-level method for appending this struct to a message.

You’re more likely to use one of the more high level functions.

fn get(&mut self, i: &mut Iter<'_>) -> Result<(), TypeMismatchError>

Low-level method for getting arguments from a message.

You’re more likely to use one of the more high level functions.

Provided Methods

fn to_emit_message(&self, path: &Path<'_>) -> Message

Returns a message that emits the signal.

fn from_message(m: &Message) -> Option<Self>

If the message is a signal of the correct type, return its arguments, otherwise return None.

This does not check sender and path of the message, which is likely relevant to you as well.

fn match_rule<'a>( sender: Option<&'a BusName<'_>>, path: Option<&'a Path<'_>> ) -> MatchRule<'a>

Returns a match rule matching this signal.

If sender and/or path is None, matches all senders and/or paths.

fn match_str(sender: Option<&BusName<'_>>, path: Option<&Path<'_>>) -> String

Returns a string that can be sent to Connection::add_match.

If sender and/or path is None, matches all senders and/or paths.