Module dbus::stdintf

This module contains some standard interfaces and an easy way to call them.

See the D-Bus specification for more information about these standard interfaces.

The code here was originally created by dbus-codegen.

Example

use dbus::{Connection, BusType};
use dbus::stdintf::org_freedesktop_dbus::Introspectable;
let c = Connection::get_private(BusType::Session).unwrap();
let p = c.with_path("org.freedesktop.DBus", "/", 10000);
println!("Introspection XML: {}", p.introspect().unwrap());

Re-exports

pub use self::org_freedesktop_dbus::Peer as OrgFreedesktopDBusPeer;

pub use self::org_freedesktop_dbus::Introspectable as OrgFreedesktopDBusIntrospectable;

pub use self::org_freedesktop_dbus::Properties as OrgFreedesktopDBusProperties;

pub use self::org_freedesktop_dbus::ObjectManager as OrgFreedesktopDBusObjectManager;

Modules

org_freedesktop_dbus