Module dbus::strings

This module contains strings with a specific format, such as a valid Interface name, a valid Error name, etc.

(The internal representation of these strings are Cow<CStr>, which makes it possible to use them in libdbus without conversion costs.)

Structs

BusName
A wrapper around a string that is guaranteed to be a valid D-Bus bus name.

ErrorName
A wrapper around a string that is guaranteed to be a valid D-Bus error name.

Interface
A wrapper around a string that is guaranteed to be a valid D-Bus interface name.

Member
A wrapper around a string that is guaranteed to be a valid D-Bus member, i e, a signal or method name.

Path
A wrapper around a string that is guaranteed to be a valid D-Bus object path.

Signature
A wrapper around a string that is guaranteed to be a valid (single) D-Bus type signature.