Struct dbus::ConnPath

pub struct ConnPath<'a, C> {
    pub conn: C,
    pub dest: BusName<'a>,
    pub path: Path<'a>,
    pub timeout: i32,
}

A convenience struct that wraps connection, destination and path.

Useful if you want to make many method calls to the same destination path.

Fields
conn: C dest: BusName<'a> path: Path<'a> timeout: i32

Timeout in milliseconds

Methods

impl<'a, C: Deref<Target=Connection>> ConnPath<'a, C>

fn method_call_with_args<F: FnOnce(&mut Message)>(&self,
i: &Interface,
m: &Member,
f: F)
-> Result<Message, Error>

Make a D-Bus method call, where you can append arguments inside the closure.

fn signal_with_args<F: FnOnce(&mut Message)>(&self,
i: &Interface,
m: &Member,
f: F)
-> Result<u32, Error>