Struct dbus::Message

pub struct Message {
    // some fields omitted
}

A D-Bus message. A message contains some headers (e g sender and destination address) and a list of MessageItems.

Methods

impl Message

fn new_method_call(destination: &str, path: &str, iface: &str, method: &str) -> Option<Message>

fn new_signal(path: &str, iface: &str, method: &str) -> Option<Message>

fn new_method_return(m: &Message) -> Option<Message>

fn new_error(m: &Message, error_name: &str, error_message: &str) -> Option<Message>

fn get_items(&self) -> Vec<MessageItem>

fn get_serial(&self) -> u32

fn append_items(&mut self, v: &[MessageItem])

fn msg_type(&self) -> MessageType

fn sender(&self) -> Option<String>

fn headers(&self) -> (MessageType, Option<String>, Option<String>, Option<String>)

Returns a tuple of (Message type, Path, Interface, Member) of the current message.

fn as_result(&mut self) -> Result<&mut Message, Error>

When the remote end returns an error, the message itself is correct but its contents is an error. This method will transform such an error to a D-Bus Error or otherwise return the original message.