Struct dbus::MethodErr

pub struct MethodErr(_, _);

A D-Bus Method Error, containing an error name and a description.

Unlike the “Error” struct, this is a Rust native struct.

Implementations

impl MethodErr

pub fn invalid_arg<T: Debug + ?Sized>(a: &T) -> MethodErr

Create an Invalid Args MethodErr.

pub fn no_arg() -> MethodErr

Create a MethodErr that there are not enough arguments given.

pub fn failed<T: Display + ?Sized>(a: &T) -> MethodErr

Create a MethodErr that the method failed in the way specified.

pub fn no_path<T: Display + ?Sized>(a: &T) -> MethodErr

Create a MethodErr that the Object path was unknown.

pub fn no_interface<T: Display + ?Sized>(a: &T) -> MethodErr

Create a MethodErr that the Interface was unknown.

pub fn no_method<T: Display + ?Sized>(a: &T) -> MethodErr

Create a MethodErr that the Method was unknown.

pub fn no_property<T: Display + ?Sized>(a: &T) -> MethodErr

Create a MethodErr that the Property was unknown.

pub fn ro_property<T: Display + ?Sized>(a: &T) -> MethodErr

Create a MethodErr that the Property was read-only.

pub fn errorname(&self) -> &ErrorName<'static>

Error name accessor

pub fn description(&self) -> &str

Description accessor

pub fn to_message(&self, msg: &Message) -> Message

Creates an error reply from a method call message.

Note: You normally don’t need to use this function, as it is called internally from Tree::handle.