Trait dbus::MsgHandler

pub trait MsgHandler {
    // Required method
    fn handler_type(&self) -> MsgHandlerType;

    // Provided method
    fn handle_msg(&mut self, _msg: &Message) -> Option<MsgHandlerResult> { ... }
}

A trait for handling incoming messages.

Required Methods

fn handler_type(&self) -> MsgHandlerType

Type of messages for which the handler will be called

Note: The return value of this function might be cached, so it must return the same value all the time.

Provided Methods

fn handle_msg(&mut self, _msg: &Message) -> Option<MsgHandlerResult>

Function to be called if the message matches the MsgHandlerType