Struct dbus::tree::leaves::Property

pub struct Property<M: MethodType<D>, D: DataType> {
    name: String,
    data: D::Property,
    sig: Signature<'static>,
    emits: EmitsChangedSignal,
    rw: Access,
    get_cb: Option<DebugGetProp<M, D>>,
    set_cb: Option<DebugSetProp<M, D>>,
    anns: Annotations,
}

A D-Bus Property.

Fields
name: String data: D::Property sig: Signature<'static> emits: EmitsChangedSignal rw: Access get_cb: Option<DebugGetProp<M, D>> set_cb: Option<DebugSetProp<M, D>> anns: Annotations
Methods

impl<M: MethodType<D>, D: DataType> Property<M, D>

fn emits_changed(self, e: EmitsChangedSignal) -> Self

Builder method that allows setting the Property's signal behavior when changed.

Note: if e is set to const, the property will be read only.

fn access(self, e: Access) -> Self

Builder method that allows setting the Property as readable, writable, or both.

Note: might modify emits_changed as well, if property is changed to non-readonly and emit is set to "Const".

fn annotate<N: Into<String>, V: Into<String>>(self, name: N, value: V) -> Self

Builder method that adds an annotation to the method.

fn deprecated(self) -> Self

Builder method that adds an annotation that this entity is deprecated.

fn get_name(&self) -> &str

Get property name

fn get_data(&self) -> &D::Property

Get associated data

fn can_get(&self) -> Result<(), MethodErr>

Returns Ok if the property is gettable

fn get_as_variant(&self,
i: &mut IterAppend,
pinfo: &PropInfo<M, D>)
-> Result<(), MethodErr>

Calls the on_get function and appends the result as a variant.

Note: Will panic if get_cb is not set.

fn can_set(&self, i: Option<Iter>) -> Result<(), MethodErr>

Returns Ok if the property is settable.

Will verify signature in case iter is not None; iter is supposed to point at the Variant with the item inside.

fn set_as_variant(&self,
i: &mut Iter,
pinfo: &PropInfo<M, D>)
-> Result<Option<Message>, MethodErr>

Calls the on_set function, which reads from i.

The return value might contain an extra message containing the EmitsChanged signal. Note: Will panic if set_cb is not set.

fn get_signal(&self, p: &PropInfo<M, D>) -> Message

Gets the signal (if any) associated with the Property.

fn get_emits_changed_signal(&self,
m: &PropInfo<M, D>)
-> Result<Option<Message>, MethodErr>

impl<'a, D: DataType> Property<MTFn<D>, D>

fn on_get<H>(self, handler: H) -> Property<MTFn<D>, D> where H: 'static + Fn(&mut IterAppend, &PropInfo<MTFn<D>, D>) -> Result<(), MethodErr>

Sets the callback for getting a property.

For single-thread use.

fn on_set<H>(self, handler: H) -> Property<MTFn<D>, D> where H: 'static + Fn(&mut Iter, &PropInfo<MTFn<D>, D>) -> Result<(), MethodErr>

Sets the callback for setting a property.

For single-thread use.

impl<'a, D: DataType> Property<MTFnMut<D>, D>

fn on_get<H>(self, handler: H) -> Property<MTFnMut<D>, D> where H: 'static + Fn(&mut IterAppend, &PropInfo<MTFnMut<D>, D>) -> Result<(), MethodErr>

Sets the callback for getting a property.

For single-thread use.

fn on_set<H>(self, handler: H) -> Property<MTFnMut<D>, D> where H: 'static + Fn(&mut Iter, &PropInfo<MTFnMut<D>, D>) -> Result<(), MethodErr>

Sets the callback for setting a property.

For single-thread use.

impl<D: DataType> Property<MTSync<D>, D>

fn on_get<H>(self, handler: H) -> Property<MTSync<D>, D> where H: Fn(&mut IterAppend, &PropInfo<MTSync<D>, D>) -> Result<(), MethodErr> + Send + Sync + 'static

Sets the callback for getting a property.

For multi-thread use.

fn on_set<H>(self, handler: H) -> Property<MTSync<D>, D> where H: Fn(&mut Iter, &PropInfo<MTSync<D>, D>) -> Result<(), MethodErr> + Send + Sync + 'static

Sets the callback for setting a property.

For single-thread use.

impl<M: MethodType<D>, D: DataType> Property<M, D> where D::Property: Append + Clone

fn default_get(self) -> Self

impl<M: MethodType<D>, D: DataType> Property<M, D> where D::Property: RefArg

fn default_get_refarg(self) -> Self