Struct dbus::tree::leaves::Signal

pub struct Signal<D: DataType> {
    name: Member<'static>,
    data: D::Signal,
    arguments: Vec<Argument>,
    anns: Annotations,
}

A D-Bus Signal.

Fields
name: Member<'static>data: D::Signalarguments: Vec<Argument>anns: Annotations
Implementations

impl<D: DataType> Signal<D>

pub fn arg<A: Into<Argument>>(self, a: A) -> Self

Builder method that adds an Argument to the Signal.

pub fn sarg<A: Arg, S: Into<String>>(self, s: S) -> Self

Builder method that adds an Argument to the Signal.

pub fn args<Z: Into<Argument>, A: IntoIterator<Item = Z>>(self, a: A) -> Self

Builder method that adds multiple Arguments to the Signal.

pub fn annotate<N: Into<String>, V: Into<String>>(
self,
name: N,
value: V
) -> Self

Add an annotation to this Signal.

pub fn deprecated(self) -> Self

Add an annotation that this entity is deprecated.

pub fn get_name(&self) -> &Member<'static>

Get signal name

pub fn get_data(&self) -> &D::Signal

Get associated data

pub fn emit(
&self,
p: &Path<'static>,
i: &IfaceName<'static>,
items: &[MessageItem]
) -> Message

Returns a message which emits the signal when sent.

Same as “msg” but also takes a “MessageItem” argument.

pub fn msg(&self, p: &Path<'static>, i: &IfaceName<'static>) -> Message

Returns a message which emits the signal when sent.

Same as “emit” but does not take a “MessageItem” argument.