Module dbus::tree::methodtype

Structs

MTFn
An abstract type to represent Fn functions.

MTFnMut
An abstract type to represent FnMut functions.

MTSync
An abstract type to represent Fn + Send + Sync functions (that can be called from several threads in parallel).

MethodErr
A D-Bus Method Error, containing an error name and a description.

MethodInfo
Contains information about the incoming method call.

PropInfo
Contains information about the incoming property get/set request.

Traits

DataType
Associated data for different objects in a tree.

MethodType
A helper trait used internally to make the tree generic over MTFn, MTFnMut and MTSync.

Functions

org_freedesktop_dbus_introspectable_server

Type Definitions

MethodResult
Result containing the Messages returned from the Method, or a MethodErr.