Struct dbus::tree::objectpath::ObjectPath

pub struct ObjectPath<M: MethodType<D>, D: DataType> {
    name: Arc<Path<'static>>,
    ifaces: BTreeMap<Arc<IfaceName<'static>>, Arc<Interface<M, D>>>,
    ifacecache: Arc<IfaceCache<M, D>>,
    data: D::ObjectPath,
}

A D-Bus Object Path.

Fields
name: Arc<Path<'static>> ifaces: BTreeMap<Arc<IfaceName<'static>>, Arc<Interface<M, D>>> ifacecache: Arc<IfaceCache<M, D>> data: D::ObjectPath
Methods

impl<M: MethodType<D>, D: DataType> ObjectPath<M, D>

fn add<I: Into<Arc<Interface<M, D>>>>(self, s: I) -> Self

Builder function that adds a interface to the object path.

fn get_name(&self) -> &Path<'static>

Get property name

fn get_data(&self) -> &D::ObjectPath

Get associated data

fn introspectable(self) -> Self

Adds introspection support for this object path.

fn object_manager(self) -> Self

Adds ObjectManager support for this object path.

It is not possible to add/remove interfaces while the object path belongs to a tree, hence no InterfacesAdded / InterfacesRemoved signals are sent.

fn introspect(&self, tree: &Tree<M, D>) -> String

fn add_property_handler(&mut self)

fn get_iface<'a>(&'a self,
iface_name: &'a CStr)
-> Result<&Arc<Interface<M, D>>, MethodErr>

fn prop_get(&self, m: &MethodInfo<M, D>) -> MethodResult

fn prop_get_all(&self, m: &MethodInfo<M, D>) -> MethodResult

fn prop_set(&self, m: &MethodInfo<M, D>) -> MethodResult

fn get_managed_objects(&self, m: &MethodInfo<M, D>) -> MethodResult

fn handle(&self, m: &Message, t: &Tree<M, D>) -> MethodResult