Struct dbus::tree::Interface

pub struct Interface<M> {
    name: Arc<IfaceName<'static>>,
    methods: BTreeMap<Arc<Member<'static>>, Arc<Method<M>>>,
    signals: BTreeMap<Arc<Member<'static>>, Arc<Signal>>,
    properties: BTreeMap<Arc<String>, Arc<Property<M>>>,
    anns: BTreeMap<String, String>,
}

Represents a D-Bus interface.

Fields
name: Arc<IfaceName<'static>> methods: BTreeMap<Arc<Member<'static>>, Arc<Method<M>>> signals: BTreeMap<Arc<Member<'static>>, Arc<Signal>> properties: BTreeMap<Arc<String>, Arc<Property<M>>> anns: BTreeMap<String, String>
Methods

impl<M> Interface<M>

fn add_m(self, m: Method<M>) -> Self

Adds a method to the interface.

fn add_s(self, s: Signal) -> Self

Adds a signal to the interface.

fn add_s_arc(self, s: Arc<Signal>) -> Self

Adds a signal to the interface. Lets you keep another clone of the signal (which you can use to emit the signal, once it belongs to an object path).

Note: You are not allowed to add a signal to more than one interface.

fn add_s_ref(&mut self, s: Signal) -> Arc<Signal>

Adds a signal to the interface. Returns a reference to the signal (which you can use to emit the signal, once it belongs to an object path).

fn add_p(self, p: Property<M>) -> Self

Adds a property to the interface.

fn add_p_arc(self, p: Arc<Property<M>>) -> Self

Adds a property to the interface. Lets you keep another clone of the property (which you can use to get and set the current value of the property).

Note: You are not allowed to add a property to more than one interface. Later function calls might panic if you do so.

fn add_p_ref(&mut self, p: Property<M>) -> Arc<Property<M>>

Adds a property to the interface. Returns a reference to the property (which you can use to get and set the current value of the property).

fn annotate<N: Into<String>, V: Into<String>>(self, name: N, value: V) -> Self

Add an annotation to this Inteface.

fn deprecated(self) -> Self

Add an annotation that this entity is deprecated.

fn new(t: IfaceName<'static>) -> Interface<M>