Struct dbus::tree::ObjectPath

pub struct ObjectPath<M: MethodType<D>, D: DataType> {
    name: Arc<Path<'static>>,
    default_iface: Option<Interface<'static>>,
    ifaces: BTreeMap<Arc<Interface<'static>>, Arc<Interface<M, D>>>,
    ifacecache: Arc<IfaceCache<M, D>>,
    data: D::ObjectPath,
}

A D-Bus Object Path.

Fields
name: Arc<Path<'static>>default_iface: Option<Interface<'static>>ifaces: BTreeMap<Arc<Interface<'static>>, Arc<Interface<M, D>>>ifacecache: Arc<IfaceCache<M, D>>data: D::ObjectPath
Implementations

impl<M: MethodType<D>, D: DataType> ObjectPath<M, D>

pub fn get_name(&self) -> &Path<'static>

Get property name

pub fn get_data(&self) -> &D::ObjectPath

Get associated data

pub fn iter<'a>(&'a self) -> Iter<'a, Interface<M, D>>

Iterates over interfaces implemented by this object path.

pub(super) fn introspect(&self, tree: &Tree<M, D>) -> String

fn get_iface<'a>( &'a self, iface_name: &'a CStr ) -> Result<&Arc<Interface<M, D>>, MethodErr>

fn prop_get(&self, m: &MethodInfo<'_, M, D>) -> MethodResult

fn prop_get_all(&self, m: &MethodInfo<'_, M, D>) -> MethodResult

fn prop_set(&self, m: &MethodInfo<'_, M, D>) -> MethodResult

fn get_managed_objects(&self, m: &MethodInfo<'_, M, D>) -> MethodResult

fn handle(&self, m: &Message, t: &Tree<M, D>) -> MethodResult

impl<M: MethodType<D>, D: DataType> ObjectPath<M, D>where <D as DataType>::Interface: Default, <D as DataType>::Method: Default,

pub fn introspectable(self) -> Self

Adds introspection support for this object path.

pub fn add<I: Into<Arc<Interface<M, D>>>>(self, s: I) -> Self

Builder function that adds a interface to the object path.

pub fn default_interface(self, i: IfaceName<'static>) -> Self

Builder function that sets what interface should be dispatched on an incoming method call without interface.

pub fn object_manager(self) -> Self

Adds ObjectManager support for this object path.

It is not possible to add/remove interfaces while the object path belongs to a tree, hence no InterfacesAdded / InterfacesRemoved signals are sent.

fn add_property_handler(&mut self)