Struct dbus::tree::Signal

pub struct Signal {
    name: Arc<Member<'static>>,
    arguments: Vec<Argument>,
    owner: Mutex<Option<(Arc<Path<'static>>, Arc<IfaceName<'static>>)>>,
    anns: BTreeMap<String, String>,
}

A D-Bus Signal.

Fields

name

arguments

owner

anns

Methods

impl Signal

fn emit(&self, items: &[MessageItem]) -> Message

Returns a message which emits the signal when sent. Panics if the signal is not inserted in an object path.

fn arg<A: Into<Argument>>(self, a: A) -> Self

Builder method that adds an Argument to the Signal.

fn args<Z: Into<Argument>, A: IntoIterator<Item=Z>>(self, a: A) -> Self

Builder method that adds multiple "rguments to the Signel.

fn annotate<N: Into<String>, V: Into<String>>(self, name: N, value: V) -> Self

Add an annotation to this Signal.

fn deprecated(self) -> Self

Add an annotation that this entity is deprecated.