Trait dbus::tree::MethodType

pub trait MethodType<D: DataType>: Sized + Default {
    type Method: ?Sized;
    type GetProp: ?Sized;
    type SetProp: ?Sized;
    fn call_getprop(_: &Self::GetProp,
                    _: &mut IterAppend,
                    _: &PropInfo<Self, D>)
                    -> Result<(), MethodErr>;
    fn call_setprop(_: &Self::SetProp,
                    _: &mut Iter,
                    _: &PropInfo<Self, D>)
                    -> Result<(), MethodErr>;
    fn call_method(_: &Self::Method, _: &MethodInfo<Self, D>) -> MethodResult;
    fn make_getprop<H>(h: H) -> Box<Self::GetProp> where H: Fn(&mut IterAppend, &PropInfo<Self, D>) -> Result<(), MethodErr> + Send + Sync + 'static;
    fn make_method<H>(h: H) -> Box<Self::Method> where H: Fn(&MethodInfo<Self, D>) -> MethodResult + Send + Sync + 'static;
}

A helper trait used internally to make the tree generic over MTFn, MTFnMut and MTSync.

You should not need to call these methods directly, it's primarily for internal use.

Associated Types

type Method: ?Sized

type GetProp: ?Sized

type SetProp: ?Sized

Required Methods

fn call_getprop(_: &Self::GetProp,
_: &mut IterAppend,
_: &PropInfo<Self, D>)
-> Result<(), MethodErr>

fn call_setprop(_: &Self::SetProp,
_: &mut Iter,
_: &PropInfo<Self, D>)
-> Result<(), MethodErr>

fn call_method(_: &Self::Method, _: &MethodInfo<Self, D>) -> MethodResult

fn make_getprop<H>(h: H) -> Box<Self::GetProp> where H: Fn(&mut IterAppend, &PropInfo<Self, D>) -> Result<(), MethodErr> + Send + Sync + 'static

fn make_method<H>(h: H) -> Box<Self::Method> where H: Fn(&MethodInfo<Self, D>) -> MethodResult + Send + Sync + 'static