Crate dbus

D-Bus bindings for Rust

D-Bus is a message bus, and is mainly used in Linux for communication between processes. It is present by default on almost every Linux distribution out there, and runs in two instances - one per session, and one system-wide.

See the examples directory for some examples to get you started.

Modules

obj
Contains functionality for the "server" of a D-Bus object. A remote application can introspect this object and call methods on it.

Structs

Connection
A D-Bus connection. Start here if you want to get on the D-Bus!

ConnectionItems
ConnectionItem iterator

Error
D-Bus Error wrapper

Message
A D-Bus message. A message contains some headers (e g sender and destination address) and a list of MessageItems.

OwnedFd
An RAII wrapper around Fd to ensure that file descriptor is closed when the scope ends.

PropHandler
Wrapper around Props that keeps a map of fetched properties.

Props
Client side properties - get and set properties on a remote application.

Enums

ArrayError

BusType

ConnectionItem
When listening for incoming events on the D-Bus, this enum will tell you what type of incoming event has happened.

MessageItem
MessageItem - used as parameters and return values from method calls, or as data added to a signal.

MessageType

NameFlag

ReleaseNameReply

RequestNameReply

Type Definitions

TypeSig
A TypeSig describes the type of a MessageItem.