Crate dbus

D-Bus bindings for Rust

D-Bus is a message bus, and is mainly used in Linux for communication between processes. It is present by default on almost every Linux distribution out there, and runs in two instances - one per session, and one system-wide.

In addition to the API documentation, which you're currently reading, you might want to look in the examples directory, which contains many examples and some additional documents. README.md also contains a few quick "getting started" examples (as well as information about the futures and no-string-validation features).

In addition to this crate, there are some companion crates:

dbus-tokio for integrating D-Bus with Tokio

dbus-codegen for generating code from D-Bus introspection data

libdbus-sys contains the raw bindings to the C libdbus library.

Re-exports

pub use crate::message::Message;

pub use crate::message::MessageType;

pub use crate::strings::Signature;

pub use crate::strings::Path;

Modules

arg
Types and traits for easily getting a message's arguments, or appening a message with arguments.

blocking
Connections and proxies that make blocking method calls.

channel
Connection base / building block.

ffidisp
A connection that uses FFI callbacks to dispatch messages.

message
Contains structs and traits closely related to D-Bus messages.

strings
This module contains strings with a specific format, such as a valid Interface name, a valid Error name, etc.

tree
Contains functionality for dispatching methods on a D-Bus "server".

Structs

Error
D-Bus Error wrapper.