Module dbus::arg

Types and traits for easily getting a message’s arguments, or appening a message with arguments.

Also see the argument’s guide (in the examples directory) for details about which Rust types correspond to which D-Bus types.

A message has read1, read2 etc, and append1, append2 etc, which is one starting point into this module’s types.

Modules

messageitem

MessageItem - old, enum design that is used as parameters and return values from method calls, or as data added to a signal.

Structs

Array

Represents a D-Bus Array. Maximum flexibility (wraps an iterator of items to append).

Dict

Append a D-Bus dict type (i e, an array of dict entries).

Iter

Helper struct for retrieve one or more arguments from a Message.

IterAppend

Helper struct for appending one or more arguments to a Message.

OwnedFd

An RAII wrapper around Fd to ensure that file descriptor is closed when the scope ends.

TypeMismatchError

Error struct to indicate a D-Bus argument type mismatch.

Variant

A simple wrapper to specify a D-Bus variant.

Enums

ArgType

Type of Argument

Traits

Append

Types that can be appended to a message as arguments implement this trait.

AppendAll

Helper trait to append many arguments to a message.

Arg

Types that can represent a D-Bus message argument implement this trait.

ArgAll

Helper trait to introspect many arguments.

DictKey

Types that can be used as keys in a dict type implement this trait.

FixedArray

If a type implements this trait, it means the size and alignment is the same as in D-Bus. This means that you can quickly append and get slices of this type.

Get

Types that can be retrieved from a message as arguments implement this trait.

ReadAll

Helper trait to read all arguments from a message.

RefArg

Object safe version of Arg + Append + Get.

Functions

cast

Cast a RefArg as a specific type (shortcut for any + downcast)

cast_mut

Cast a RefArg as a specific type (shortcut for any_mut + downcast_mut)

prop_cast

Descend into a hashmap returned by e g “Properties::get_all” to retrieve the value of a property.

Type Definitions

PropMap

The type typically used for a dictionary of properties.