Module dbus::arg::messageitem

MessageItem - old, enum design that is used as parameters and return values from method calls, or as data added to a signal.

Note that the newer generic design (see arg module) is, in general, both faster and smaller than MessageItem, and should be your first hand choice whenever applicable. There is also a trait object design called RefArg in case the generic design is too inflexible.

Structs

MessageItemArray
An array of MessageItem where every MessageItem is of the same type.

MessageItemDict
An array of MessageItem where every MessageItem is of the same type.

PropHandler
Wrapper around Props that keeps a map of fetched properties.

Props
Client side properties - get and set properties on a remote application.

Enums

ArrayError
Errors that can happen when creating a MessageItem::Array.

MessageItem
MessageItem - used as parameters and return values from method calls, or as data added to a signal (old, enum version).