Module dbus::arg::msgarg

Traits

Append
Types that can be appended to a message as arguments implement this trait.

Arg
Types that can represent a D-Bus message argument implement this trait.

DictKey
Types that can be used as keys in a dict type implement this trait.

FixedArray
If a type implements this trait, it means the size and alignment is the same as in D-Bus. This means that you can quickly append and get slices of this type.

Get
Types that can be retrieved from a message as arguments implement this trait.

RefArg
Object safe version of Arg + Append + Get.

Functions

cast
Cast a RefArg as a specific type (shortcut for any + downcast)

cast_mut
Cast a RefArg as a specific type (shortcut for any_mut + downcast_mut)