Module dbus::arg::msgarg

Traits

Append
Types that can be appended to a message as arguments implement this trait.

AppendAll
Helper trait to append many arguments to a message.

Arg
Types that can represent a D-Bus message argument implement this trait.

ArgBuilder
Internal trait to help generics. Implemented for (), (A1), (A1, A2) and so on (where A1: Arg, A2: Arg etc).

DictKey
Types that can be used as keys in a dict type implement this trait.

FixedArray
If a type implements this trait, it means the size and alignment is the same as in D-Bus. This means that you can quickly append and get slices of this type.

Get
Types that can be retrieved from a message as arguments implement this trait.

ReadAll
Helper trait to read all arguments from a message.

RefArg
Object safe version of Arg + Append + Get.

Functions

cast
Cast a RefArg as a specific type (shortcut for any + downcast)

cast_mut
Cast a RefArg as a specific type (shortcut for any_mut + downcast_mut)