Trait dbus::arg::msgarg::RefArg

pub trait RefArg: Debug {
    fn arg_type(&self) -> ArgType;
    fn signature(&self) -> Signature<'static>;
    fn append(&self, _: &mut IterAppend);
    fn as_any(&self) -> &Any where Self: 'static;
    fn as_any_mut(&mut self) -> &mut Any where Self: 'static;

    fn as_i64(&self) -> Option<i64> { ... }
    fn as_str(&self) -> Option<&str> { ... }
    fn as_iter<'a>(&'a self) -> Option<Box<Iterator<Item=&'a RefArg> + 'a>> { ... }
}

Object safe version of Arg + Append + Get.

This trait is somewhat under development, which means that not all types are supported and that the API might change. Only use in case Arg is not dynamic enough for your needs.

Required Methods

fn arg_type(&self) -> ArgType

The corresponding D-Bus argument type code.

fn signature(&self) -> Signature<'static>

The corresponding D-Bus type signature for this type.

fn append(&self, _: &mut IterAppend)

Performs the append operation.

fn as_any(&self) -> &Any where Self: 'static

Transforms this argument to Any (which can be downcasted to read the current value).

fn as_any_mut(&mut self) -> &mut Any where Self: 'static

Transforms this argument to Any (which can be downcasted to read the current value).

Panic

Will panic if the interior cannot be made mutable, e g, if encapsulated inside a Rc with a reference count > 1.

Provided Methods

fn as_i64(&self) -> Option<i64>

Try to read the argument as an i64.

fn as_str(&self) -> Option<&str>

Try to read the argument as a str.

fn as_iter<'a>(&'a self) -> Option<Box<Iterator<Item=&'a RefArg> + 'a>>

Try to read the argument as an iterator.