Struct dbus::arg::OwnedFd

pub struct OwnedFd { /* fields omitted */ }

An RAII wrapper around Fd to ensure that file descriptor is closed when the scope ends.

Implementations

impl OwnedFd

pub unsafe fn new(fd: RawFd) -> OwnedFd

Create a new OwnedFd from a RawFd.

This function is unsafe, because you could potentially send in an invalid file descriptor, or close it during the lifetime of this struct. This could potentially be unsound.

pub fn into_fd(self) -> RawFd

Convert an OwnedFD back into a RawFd.