Trait dbus::arg::RefArg

pub trait RefArg: Debug {
    fn arg_type(&self) -> ArgType;

    fn signature(&self) -> Signature<'static>;

    fn append(&self, _: &mut IterAppend);

    fn as_any(&self) -> &Any
    where
        Self: 'static;

    fn as_any_mut(&mut self) -> &mut Any
    where
        Self: 'static;

    fn as_i64(&self) -> Option<i64> { ... }

    fn as_u64(&self) -> Option<u64> { ... }

    fn as_f64(&self) -> Option<f64> { ... }

    fn as_str(&self) -> Option<&str> { ... }

    fn as_iter<'a>(&'a self) -> Option<Box<Iterator<Item = &'a RefArg> + 'a>> { ... }

    fn box_clone(&self) -> Box<RefArg + 'static> { ... }
}

Object safe version of Arg + Append + Get.

Required Methods

fn arg_type(&self) -> ArgType

The corresponding D-Bus argument type code.

fn signature(&self) -> Signature<'static>

The corresponding D-Bus type signature for this type.

fn append(&self, _: &mut IterAppend)

Performs the append operation.

fn as_any(&self) -> &Any where
Self: 'static,

Transforms this argument to Any (which can be downcasted to read the current value).

Note: The internal representation of complex types (Array, Dict, Struct) is unstable and as_any should not be relied upon for these types. Use as_iter instead.

fn as_any_mut(&mut self) -> &mut Any where
Self: 'static,

Transforms this argument to Any (which can be downcasted to read the current value).

Note: The internal representation of complex types (Array, Dict, Struct) is unstable and as_any should not be relied upon for these types. Use as_iter instead.

Panic

Will panic if the interior cannot be made mutable, e g, if encapsulated inside a Rc with a reference count > 1.

Provided Methods

fn as_i64(&self) -> Option<i64>

Try to read the argument as an i64.

Works for: Boolean, Byte, Int16, UInt16, Int32, UInt32, Int64, UnixFd.

fn as_u64(&self) -> Option<u64>

Try to read the argument as an u64.

Works for: Boolean, Byte, Int16, UInt16, Int32, UInt32, UInt64.

fn as_f64(&self) -> Option<f64>

Try to read the argument as an f64.

Works for: Boolean, Byte, Int16, UInt16, Int32, UInt32, Double.

fn as_str(&self) -> Option<&str>

Try to read the argument as a str.

Works for: String, ObjectPath, Signature.

fn as_iter<'a>(&'a self) -> Option<Box<Iterator<Item = &'a RefArg> + 'a>>

Try to read the argument as an iterator.

Works for: Array/Dict, Struct, Variant.

Important traits for Box<R>
impl<R> Read for Box<R> where
R: Read + ?Sized, impl<W> Write for Box<W> where
W: Write + ?Sized, impl<I> Iterator for Box<I> where
I: Iterator + ?Sized, type Item = <I as Iterator>::Item;

fn box_clone(&self) -> Box<RefArg + 'static>

Deep clone of the RefArg, causing the result to be 'static.

Usable as an escape hatch in case of lifetime problems with RefArg.

In case of complex types (Array, Dict, Struct), the clone is not guaranteed to have the same internal representation as the original.