Trait dbus::blocking::BlockingSender

pub trait BlockingSender {
    fn send_with_reply_and_block(
        &self, 
        msg: Message, 
        timeout: Duration
    ) -> Result<Message, Error>;
}

Abstraction over different connections

Required methods

fn send_with_reply_and_block(
&self,
msg: Message,
timeout: Duration
) -> Result<Message, Error>

Sends a message over the D-Bus and blocks, waiting for a reply or a timeout. This is used for method calls.

Note: In case of an error reply, this is returned as an Err(), not as a Ok(Message) with the error type.

Loading content...