Module dbus::connection

Structs

ConnMsgs
Iterator over incoming messages on a connection.

Connection
A D-Bus connection. Start here if you want to get on the D-Bus!

ConnectionItems
ConnectionItem iterator

IConnection

MessageReply
The struct returned from Connection::send_and_reply.

MsgHandlerResult
The result from MsgHandler::handle.

Enums

ConnectionItem
When listening for incoming events on the D-Bus, this enum will tell you what type of incoming event has happened.

DBusNameFlag
Flags to use for Connection::register_name.

MsgHandlerType
Type of messages to be handled by a MsgHandler.

Traits

MsgHandler
A trait for handling incoming messages.

Functions

conn_handle

default_filter_callback

filter_message_cb

msghandler_process

object_path_message_cb

Type Definitions

MessageCallback
The type of function to use for replacing the message callback.

MsgHandlerList