Enum dbus::MessageItem

pub enum MessageItem {
    Array(MessageItemArray),
    Struct(Vec<MessageItem>),
    Variant(Box<MessageItem>),
    DictEntry(Box<MessageItem>, Box<MessageItem>),
    ObjectPath(Path<'static>),
    Str(String),
    Bool(bool),
    Byte(u8),
    Int16(i16),
    Int32(i32),
    Int64(i64),
    UInt16(u16),
    UInt32(u32),
    UInt64(u64),
    Double(f64),
    UnixFd(OwnedFd),
}

MessageItem - used as parameters and return values from method calls, or as data added to a signal (old, enum version).

Note that the newer generic design (see arg module) is both faster and less error prone than MessageItem, and should be your first hand choice whenever applicable.

Variants
Array(MessageItemArray)

A D-Bus array requires all elements to be of the same type. All elements must match the Signature.

Struct(Vec<MessageItem>)

A D-Bus struct allows for values of different types.

Variant(Box<MessageItem>)

A D-Bus variant is a wrapper around another MessageItem, which can be of any type.

DictEntry(Box<MessageItem>, Box<MessageItem>)

A D-Bus dictionary entry. These are only allowed inside an array.

ObjectPath(Path<'static>)

A D-Bus objectpath requires its content to be a valid objectpath, so this cannot be any string.

Str(String)

A D-Bus String is zero terminated, so no \0 s in the String, please. (D-Bus strings are also - like Rust strings - required to be valid UTF-8.)

Bool(bool)

A D-Bus boolean type.

Byte(u8)

A D-Bus unsigned 8 bit type.

Int16(i16)

A D-Bus signed 16 bit type.

Int32(i32)

A D-Bus signed 32 bit type.

Int64(i64)

A D-Bus signed 64 bit type.

UInt16(u16)

A D-Bus unsigned 16 bit type.

UInt32(u32)

A D-Bus unsigned 32 bit type.

UInt64(u64)

A D-Bus unsigned 64 bit type.

Double(f64)

A D-Bus IEEE-754 double-precision floating point type.

UnixFd(OwnedFd)

D-Bus allows for sending file descriptors, which can be used to set up SHM, unix pipes, or other communication channels.

Methods

impl MessageItem

pub fn signature(&self) -> Signature<'static>

Get the D-Bus Signature for this MessageItem.

Note: Since dictionary entries have no valid signature, calling this function for a dict entry will cause a panic.

pub fn type_sig(&self) -> TypeSig<'static>

Deprecated:
superseded by signature

Get the D-Bus ASCII type-code for this MessageItem.

pub fn array_type(&self) -> i32

Get the integer value for this MessageItem's type-code.

pub fn from_dict<E, I: Iterator<Item = Result<(String, MessageItem), E>>>(
i: I
) -> Result<MessageItem, E>

Creates a (String, Variant) dictionary from an iterator with Result passthrough (an Err will abort and return that Err)

pub fn new_array(v: Vec<MessageItem>) -> Result<MessageItem, ArrayError>

Creates an MessageItem::Array from a list of MessageItems.

Note: This requires v to be non-empty. See also MessageItem::from(&[T]), which can handle empty arrays as well.

pub fn inner<'a, T: FromMessageItem<'a>>(&'a self) -> Result<T, ()>

Conveniently get the inner value of a MessageItem

Example

use dbus::MessageItem;
let m: MessageItem = 5i64.into();
let s: i64 = m.inner().unwrap();
assert_eq!(s, 5i64);