Enum dbus::MessageItem

pub enum MessageItem {
    Array(Vec<MessageItem>, TypeSig<'static>),
    Struct(Vec<MessageItem>),
    Variant(Box<MessageItem>),
    DictEntry(Box<MessageItem>, Box<MessageItem>),
    ObjectPath(String),
    Str(String),
    Bool(bool),
    Byte(u8),
    Int16(i16),
    Int32(i32),
    Int64(i64),
    UInt16(u16),
    UInt32(u32),
    UInt64(u64),
    Double(f64),
    UnixFd(OwnedFd),
}

MessageItem - used as parameters and return values from method calls, or as data added to a signal.

Variants

Array

Struct

Variant

DictEntry

ObjectPath

Str

Bool

Byte

Int16

Int32

Int64

UInt16

UInt32

UInt64

Double

UnixFd

Methods

impl MessageItem

fn type_sig(&self) -> TypeSig<'static>

fn array_type(&self) -> i32

fn from_dict<E, I: Iterator<Item=Result<(String, MessageItem), E>>>(i: I) -> Result<MessageItem, E>

Creates an Array from an iterator with Result passthrough (an Err will abort and return that Err)

fn new_array(v: Vec<MessageItem>) -> Result<MessageItem, ArrayError>

Creates an MessageItem::Array from a list of MessageItems.

fn inner<'a, T: FromMessageItem<'a>>(&'a self) -> Result<T, ()>