Trait dbus::ffidisp::MsgHandler

pub trait MsgHandler {
    fn handler_type(&self) -> MsgHandlerType;

    fn handle_msg(&mut self, _msg: &Message) -> Option<MsgHandlerResult> { ... }
}

A trait for handling incoming messages.

Required methods

fn handler_type(&self) -> MsgHandlerType

Type of messages for which the handler will be called

Note: The return value of this function might be cached, so it must return the same value all the time.

Provided methods

fn handle_msg(&mut self, _msg: &Message) -> Option<MsgHandlerResult>

Function to be called if the message matches the MsgHandlerType