Module dbus::message

Structs

Message
A D-Bus message. A message contains some headers (e g sender and destination address) and a list of MessageItems.

OwnedFd
An RAII wrapper around Fd to ensure that file descriptor is closed when the scope ends.

Enums

ArrayError
Errors that can happen when creating a MessageItem::Array.

MessageItem
MessageItem - used as parameters and return values from method calls, or as data added to a signal.

Traits

FromMessageItem
Helper trait for MessageItem::inner()

Functions

append_messageitem

get_message_ptr

get_messageitem

iter_append_array

iter_append_dict

iter_append_f64

iter_append_struct

iter_append_variant

iter_get_basic

iter_get_f64

message_from_ptr

new_dbus_message_iter