Module dbus::message

Structs

ConnPath

A convenience struct that wraps connection, destination and path.

Message

A D-Bus message. A message contains some headers (e g sender and destination address) and a list of MessageItems.

MessageItemArray

An array of MessageItem where every MessageItem is of the same type.

OwnedFd

An RAII wrapper around Fd to ensure that file descriptor is closed when the scope ends.

Enums

ArrayError

Errors that can happen when creating a MessageItem::Array.

MessageItem

MessageItem - used as parameters and return values from method calls, or as data added to a signal (old, enum version).

Traits

FromMessageItem

Helper trait for MessageItem::inner()

Functions

append_messageitem

get_messageitem

iter_append_array

iter_append_dict

iter_append_struct

iter_append_variant

iter_get_basic

new_dbus_message_iter