Enum dbus::message::MessageItem

pub enum MessageItem {

    Array(MessageItemArray),
    Struct(Vec<MessageItem>),
    Variant(Box<MessageItem>),
    DictEntry(Box<MessageItem>, Box<MessageItem>),
    ObjectPath(Path<'static>),
    Str(String),
    Bool(bool),
    Byte(u8),
    Int16(i16),
    Int32(i32),
    Int64(i64),
    UInt16(u16),
    UInt32(u32),
    UInt64(u64),
    Double(f64),
    UnixFd(OwnedFd),

}

MessageItem - used as parameters and return values from method calls, or as data added to a signal (old, enum version).

Note that the newer generic design (see arg module) is both faster and less error prone than MessageItem, and should be your first hand choice whenever applicable.

Variants

Array

A D-Bus array requires all elements to be of the same type. All elements must match the Signature.

Tuple Fields of Array

0: MessageItemArray

Struct

A D-Bus struct allows for values of different types.

Tuple Fields of Struct

0: Vec<MessageItem>

Variant

A D-Bus variant is a wrapper around another MessageItem, which can be of any type.

Tuple Fields of Variant

0: Box<MessageItem>

DictEntry

A D-Bus dictionary entry. These are only allowed inside an array.

Tuple Fields of DictEntry

0: Box<MessageItem>1: Box<MessageItem>

ObjectPath

A D-Bus objectpath requires its content to be a valid objectpath, so this cannot be any string.

Tuple Fields of ObjectPath

0: Path<'static>

Str

A D-Bus String is zero terminated, so no \0 s in the String, please. (D-Bus strings are also - like Rust strings - required to be valid UTF-8.)

Tuple Fields of Str

0: String

Bool

A D-Bus boolean type.

Tuple Fields of Bool

0: bool

Byte

A D-Bus unsigned 8 bit type.

Tuple Fields of Byte

0: u8

Int16

A D-Bus signed 16 bit type.

Tuple Fields of Int16

0: i16

Int32

A D-Bus signed 32 bit type.

Tuple Fields of Int32

0: i32

Int64

A D-Bus signed 64 bit type.

Tuple Fields of Int64

0: i64

UInt16

A D-Bus unsigned 16 bit type.

Tuple Fields of UInt16

0: u16

UInt32

A D-Bus unsigned 32 bit type.

Tuple Fields of UInt32

0: u32

UInt64

A D-Bus unsigned 64 bit type.

Tuple Fields of UInt64

0: u64

Double

A D-Bus IEEE-754 double-precision floating point type.

Tuple Fields of Double

0: f64

UnixFd

D-Bus allows for sending file descriptors, which can be used to set up SHM, unix pipes, or other communication channels.

Tuple Fields of UnixFd

0: OwnedFd

Implementations

impl MessageItem

pub fn signature(&self) -> Signature<'static>

Get the D-Bus Signature for this MessageItem.

Note: Since dictionary entries have no valid signature, calling this function for a dict entry will cause a panic.

pub fn type_sig(&self) -> TypeSig<'static>

👎 Deprecated:
superseded by signature

Get the D-Bus ASCII type-code for this MessageItem.

pub fn array_type(&self) -> i32

Get the integer value for this MessageItem’s type-code.

pub fn from_dict<E, I: Iterator<Item = Result<(String, MessageItem), E>>>(
i: I
) -> Result<MessageItem, E>

Creates a (String, Variant) dictionary from an iterator with Result passthrough (an Err will abort and return that Err)

pub fn new_array(v: Vec<MessageItem>) -> Result<MessageItem, ArrayError>

Creates an MessageItem::Array from a list of MessageItems.

Note: This requires v to be non-empty. See also MessageItem::from(&[T]), which can handle empty arrays as well.

fn new_array2<D, I>(i: I) -> MessageItem where
D: Into<MessageItem>,
D: Default,
I: Iterator<Item = D>,

fn new_array3<'b, D: 'b, I>(i: I) -> MessageItem where
D: Into<MessageItem> + Default + Clone,
I: Iterator<Item = &'b D>,

fn from_iter_single(i: &mut DBusMessageIter) -> Option<MessageItem>

fn from_iter(i: &mut DBusMessageIter) -> Vec<MessageItem>

fn iter_append_basic<T>(&self, i: &mut DBusMessageIter, v: T)

fn iter_append(&self, i: &mut DBusMessageIter)

fn copy_to_iter(i: &mut DBusMessageIter, v: &[MessageItem])

pub fn inner<'a, T: FromMessageItem<'a>>(&'a self) -> Result<T, ()>

Conveniently get the inner value of a MessageItem

Example

use dbus::MessageItem;
let m: MessageItem = 5i64.into();
let s: i64 = m.inner().unwrap();
assert_eq!(s, 5i64);