Module dbus::obj

Contains functionality for the "server" of a D-Bus object. A remote application can introspect this object and call methods on it.

Structs

Argument
a Method has a list of Arguments.

Interface
Interfaces can contain Methods, Properties, and Signals.

Method
a method that can be called from another application

ObjectPath
Represents a D-Bus object path, which can in turn contain Interfaces.

Property
Properties that a remote application can get/set.

Signal
Declares that an Interface can send this signal

Traits

PropertyROHandler

PropertyRWHandler

PropertyWOHandler

Type Definitions

MethodHandler
A boxed closure for dynamic dispatch. It is called when the method is called by a remote application.

MethodResult
A method returns either a list of MessageItems, or an error - the tuple represents the name and message of the Error.

PropertyGetResult

PropertySetResult