Module dbus::stdintf

This module contains some standard interfaces and an easy way to call them.

See the D-Bus specification for more information about these standard interfaces.

The code here was originally created by dbus-codegen. Signal handling is not yet implemented here - use Message::signal to emit or Connection.add_match to receive, as usual.

Example

use dbus::{Connection, BusType};
use dbus::stdintf::OrgFreedesktopDBusIntrospectable;
let c = Connection::get_private(BusType::Session).unwrap();
let p = c.with_path("org.freedesktop.DBus", "/", 10000);
println!("Introspection XML: {}", p.introspect().unwrap());

Traits

OrgFreedesktopDBusIntrospectable
Method of the org.freedesktop.DBus.Introspectable interface.

OrgFreedesktopDBusObjectManager
Method of the org.freedesktop.DBus.ObjectManager interface.

OrgFreedesktopDBusPeer
Methods of the org.freedesktop.DBus.Peer interface.

OrgFreedesktopDBusProperties
Methods of the org.freedesktop.DBus.Properties interface.