Struct dbus::tree::Interface

pub struct Interface<M> {
    // some fields omitted
}

Represents a D-Bus interface.

Methods

impl<M> Interface<M>

fn add_m(self, m: Method<M>) -> Self

Adds a method to the interface.

fn add_s(self, s: Signal) -> Self

Adds a signal to the interface.

fn add_s_ref(&mut self, s: Signal) -> Arc<Signal>

Adds a signal to the interface. Returns a reference to the signal (which you can use to emit the signal, once it belongs to an object path).

fn add_p(self, p: Property<M>) -> Self

Adds a property to the interface.

fn add_p_ref(&mut self, p: Property<M>) -> Arc<Property<M>>

Adds a property to the interface. Returns a reference to the property (which you can use to get and set the current value of the property).

fn annotate<N: Into<String>, V: Into<String>>(self, name: N, value: V) -> Self

Add an annotation to this Inteface.

fn deprecated(self) -> Self

Add an annotation that this entity is deprecated.