Struct dbus::tree::Method

pub struct Method<M: MethodType<D>, D: DataType> {
    cb: DebugMethod<M, D>,
    data: D::Method,
    name: Member<'static>,
    i_args: Vec<Argument>,
    o_args: Vec<Argument>,
    anns: Annotations,
}

A D-Bus Method.

Fields
cb: DebugMethod<M, D>data: D::Methodname: Member<'static>i_args: Vec<Argument>o_args: Vec<Argument>anns: Annotations
Implementations

impl<M: MethodType<D>, D: DataType> Method<M, D>

pub fn in_arg<A: Into<Argument>>(self, a: A) -> Self

Builder method that adds an “in” Argument to this Method.

pub fn inarg<A: Arg, S: Into<String>>(self, s: S) -> Self

Builder method that adds an “in” Argument to this Method.

pub fn in_args<Z: Into<Argument>, A: IntoIterator<Item = Z>>(self, a: A) -> Self

Builder method that adds multiple “in” Arguments to this Method.

pub fn out_arg<A: Into<Argument>>(self, a: A) -> Self

Builder method that adds an “out” Argument to this Method.

pub fn outarg<A: Arg, S: Into<String>>(self, s: S) -> Self

Builder method that adds an “out” Argument to this Method.

pub fn out_args<Z: Into<Argument>, A: IntoIterator<Item = Z>>(
self,
a: A
) -> Self

Builder method that adds multiple “out” Arguments to this Method.

pub fn annotate<N: Into<String>, V: Into<String>>(
self,
name: N,
value: V
) -> Self

Builder method that adds an annotation to the method.

pub fn deprecated(self) -> Self

Builder method that adds an annotation that this entity is deprecated.

pub fn call(&self, minfo: &MethodInfo<'_, M, D>) -> MethodResult

Call the Method

pub fn get_name(&self) -> &Member<'static>

Get method name

pub fn get_data(&self) -> &D::Method

Get associated data