Struct dbus::tree::Tree

pub struct Tree<M> {
    // some fields omitted
}

A collection of object paths.

Methods

impl<M: MethodType> Tree<M>

fn add(self, p: ObjectPath<M>) -> Self

Add a Object Path to this Tree.

fn set_registered(&self, c: &Connection, b: bool) -> Result<(), Error>

Registers or unregisters all object paths in the tree.

fn handle(&self, m: &Message) -> Option<Vec<Message>>

Handles a message. Will return None in case the object path was not found, or otherwise a list of messages to be sent back.

fn run<'a, I: Iterator<Item=ConnectionItem>>(&'a self, c: &'a Connection, i: I) -> TreeServer<'a, I, M>

This method takes an ConnectionItem iterator (you get it from Connection::iter()) and handles all matching items. Non-matching items (e g signals) are passed through.